fix(init-local): Ignore changes to code made by lerna. Control failure status.

+++ scripts/tasks.sh
@@ -5,6 +5,7 @@ BASEDIR="$(pwd)/scripts"
 fn_local_init () {
   set -e
   TEMP_PATH=${1:-"."}
+  STATUS="running"
 
   echo ">>> path: $TEMP_PATH"
   echo ">>> basedir: $BASEDIR"
@@ -19,16 +20,29 @@ fn_local_init () {
   verdaccio --config "${BASEDIR}/verdaccio_config.yaml" &
   sleep 5
 
-  echo ">>> STEP 3: lerna publish"
-  npx lerna publish prerelease --registry http://localhost:4873/ --no-git-tag-version --no-private --no-push --yes --no-git-reset --dist-tag local
+  echo ">>> STEP 3: lerna publish. Ignore any version bumps using intermediate branch."
+  git stash
+  git branch -d verdaccio-temp || true
+  git checkout -b verdaccio-temp
+  git stash apply
+  npx lerna publish prerelease --registry http://localhost:4873/ --no-git-tag-version --no-private --no-push --yes --no-git-reset --dist-tag local || STATUS="failed-lerna"
+  git reset --hard HEAD
+  git checkout -
+  git stash pop
+  git branch -d verdaccio-temp
 
-  echo ">>> STEP 4: init app"
-  rm -rf "${TEMP_PATH}/testapp"
-  npm_config_registry=http://localhost:4873/ LOCAL_DIR=${TEMP_PATH} npx startupjs@local init testapp
+  if [ "$STATUS" = "running" ]; then
+    echo ">>> STEP 4: init app"
+    rm -rf "${TEMP_PATH}/testapp"
+    npm_config_registry=http://localhost:4873/ LOCAL_DIR=${TEMP_PATH} npx startupjs@local init testapp || STATUS="failed-init"
+  fi;
 
-  echo ">>> STEP 6: kill verdaccio"
+  echo ">>> STEP 5: kill verdaccio"
   verdaccio_pid=$(lsof -i:4873 | awk 'FNR > 1 {print $2}') && [ -z "$verdaccio_pid" ] || kill $verdaccio_pid
 
-  echo ">>> STEP 7: uninstall verdaccio"
+  echo ">>> STEP 6: uninstall verdaccio"
   npm uninstall -g verdaccio verdaccio-auth-memory verdaccio-memory
+
+  if [ "$STATUS" = "failed-lerna" ]; then echo "!!! ERROR !!! 'lerna publish' failed!" && exit 1; fi
+  if [ "$STATUS" = "failed-init" ]; then echo "!!! ERROR !!! 'npx startupjs init' failed!" && exit 1; fi
 }
